Move Jax beam search decoding code
